refactor(pm): consolidate CLI enums into util/cli_enum.rs
Rename save_type.rs -> cli_enum.rs and consolidate all CLI parameter
enums (SaveType, PackageAction, OmitType, ScriptPolicy, ConfigScope,
RunMode) into one module. Replaces hand-written bool_enum! macro with
explicit hand-written enums per community best practice.

Delete bool_enum.rs. Move ConfigScope from config_file.rs and RunMode
from model/mod.rs, keeping re-exports for backward compatibility.

Module doc references:
- https://blakesmith.me/2019/05/07/rust-patterns-enums-instead-of-booleans.html
- https://github.com/rust-lang/book/issues/2186
